+++ data/src/config/pane.rs
@@ -1,0 +1,16 @@
+use serde::Deserialize;
+
+#[derive(Debug, Clone, Deserialize, Default)]
+pub struct Pane {
+    /// Default axis used when splitting a pane.
+    #[serde(default)]
+    pub split_axis: SplitAxis,
+}
+
+#[derive(Debug, Copy, Clone, Deserialize, Default)]
+#[serde(rename_all = "kebab-case")]
+pub enum SplitAxis {
+    #[default]
+    Horizontal,
+    Vertical,
+}